change about BMEP gui

+++ src/main/java/net/epitap/degeneracycraft/blocks/machine/basic/basic_power_composite_structure_type_thermal_generator/BasicPowerCompositeStructureTypeThermalGeneratorMenu.java
@@ -46,9 +46,9 @@ public class BasicPowerCompositeStructureTypeThermalGeneratorMenu extends Abstra
     }
 
 
-    public int getBurnTime() {
+    public float getBurnTime() {
         int getBurnTime = this.data.get(0);
-        return getBurnTime / 20;
+        return getBurnTime / 20F;
     }
 
     public DCIEnergyStorageFloat getEnergy() {
